test against django 4.2, drop older versions

+++ pybb/views.py
@@ -458,7 +458,8 @@ class PostEditMixin(PybbFormsMixin):
                         success = False
                 else:
                     topic.poll_question = None
-                    topic.poll_answers.all().delete()
+                    if topic.pk:
+                        topic.poll_answers.all().delete()
         else:
             pollformset = None
 
@@ -524,7 +525,7 @@ class AddPostView(PostEditMixin, generic.CreateView):
                     profile = util.get_pybb_profile(post.user)
                     self.quote = util._get_markup_quoter(defaults.PYBB_MARKUP)(post.body, profile.get_display_name())
 
-                if self.quote and request.is_ajax():
+                if self.quote and request.headers.get('x-requested-with') == 'XMLHttpRequest':
                     return HttpResponse(self.quote)
         return super(AddPostView, self).dispatch(request, *args, **kwargs)
 
@@ -748,11 +749,13 @@ class DeletePostView(generic.DeleteView):
         self.forum = post.topic.forum
         return post
 
-    def delete(self, request, *args, **kwargs):
-        self.object = self.get_object()
+    def form_valid(self, form):
+        """
+        get_success_url must be called after deletion
+        """
         self.object.delete()
         redirect_url = self.get_success_url()
-        if not request.is_ajax():
+        if not self.request.headers.get('x-requested-with') == 'XMLHttpRequest':
             return HttpResponseRedirect(redirect_url)
         else:
             return HttpResponse(redirect_url)
@@ -763,7 +766,7 @@ class DeletePostView(generic.DeleteView):
         except Topic.DoesNotExist:
             return self.forum.get_absolute_url()
         else:
-            if not self.request.is_ajax():
+            if not self.request.headers.get('x-requested-with') == 'XMLHttpRequest':
                 return self.topic.get_absolute_url()
             else:
                 return ""
